Adding cancel appointment backend
